Removed jsfour-core
I've decided to remove the jsfour-core dependency from this resource since it's not really a lot of code in it that this resource uses. Which makes it easier to install. I've also decided to remove some of the tablet code since I've put that on hold. May be re-added in the future.